Chore: separated ArtistRecommendation out into it's own component form Tab 1
