+++ src/pages/ArtistRecommendations.tsx
@@ -1,131 +1,167 @@
+import {
+  IonCard,
+  IonCardHeader,
+  IonCardSubtitle,
+  IonCardTitle,
+  IonItemGroup,
+  IonButton,
+} from '@ionic/react';
+
 import React from 'react';
-import { IonContent, IonHeader, IonPage, IonTitle, IonToolbar, IonButton, IonCard } from '@ionic/react';
-import './BookerSignup2.css';
-import { connect } from 'react-redux'
-import PlacesAutocomplete, {
-  geocodeByAddress,
-  getLatLng
-} from 'react-places-autocomplete'
-import { updatedVenue } from '../store/booker'
+import { auth, me } from '../store/user';
+import { searchBarValue } from '../store/filter';
+import { connect } from 'react-redux';
+import { getRecommendedVenues } from '../store/venue';
+import { getOneBooker } from '../store/booker';
+import { getRocommendedArtists } from '../store/artist';
+import './Tab1.css';
 
 interface IMyComponentState {
-  latitude: Number,
-  longitude: Number,
-  address: String,
-  on: boolean
+  isSearchBarOpen: boolean;
+  currentVenue: number;
+  currentBookerRecommandations: Array<object>;
 }
 interface IMyComponentProps {
-  booker: object,
-  updateVenue: any
-
+  auth: any;
+  error: any;
+  user: object;
+  me: any;
+  searchBarValue: (value: boolean) => void;
+  isSearchBarOpen: boolean;
+  getRecommendedVenues: any;
+  fetchVenues: any;
+  getRocommendedArtists: any;
+  venues: any;
+  artists: any;
+  recommendations: Array<object>;
 }
-
-// function getHaversineDistance(latA,longA,latB,longB) {
-
-// }
-
-class Login extends React.Component<IMyComponentProps, IMyComponentState> {
+class ArtistRecommendation extends React.Component<IMyComponentProps, IMyComponentState> {
   constructor(props) {
     super(props);
     this.state = {
-      latitude: 0,
-      longitude: 0,
-      address: '',
-      on: false
-    }
+      isSearchBarOpen: this.props.isSearchBarOpen,
+      currentVenue: 1,
+      currentBookerRecommandations: [],
+    };
   }
-  componentDidMount() {
-    let booker = window.localStorage.getItem('booker')
-    booker = JSON.parse(booker || '');
-    let newBooker = booker || {};
-    if (newBooker['venue'] !== undefined) {
-      this.setState({
-        address: newBooker['venue'].address,
-        latitude: newBooker['venue'].latitude,
-        longitude: newBooker['venue'].longitude
-      })
-    }
-  }
-  handleChange = address => {
-    this.setState({ address });
-  };
-
-  handleSelect = address => {
-    geocodeByAddress(address)
-      .then(results => getLatLng(results[0]))
-      .then(latLng => {
+  async componentDidMount() {
+    await this.props.me();
+    if (this.props.user['id'] !== undefined) {
+      if (this.props.user['status'] === 'booker') {
+        const id = this.props.user['id'];
+        await this.props.fetchVenues(id);
+        await this.setState({ currentVenue: this.props.venues[0].id });
+        await this.props.getRocommendedArtists(this.state.currentVenue);
+        const rec = this.props.artists.filter(
+          artist => artist['recommendations'][0].score <= 9
+        );
         this.setState({
-          latitude: latLng.lat,
-          longitude: latLng.lng,
-          address
-        })
-        console.log('Success grabbing adress!')
-      })
-      .catch(error => console.error('Error', error));
+          currentBookerRecommandations: rec,
+        });
+      }
+    }
+    let searchbar = window.localStorage.getItem('searchbar');
+    if (searchbar !== null) {
+      let value: boolean;
+      value = JSON.parse(searchbar || '');
+      this.props.searchBarValue(value);
+    }
+    this.setState({
+      isSearchBarOpen: this.props.isSearchBarOpen,
+    });
+  }
+
+  handleChange = async e => {
+    this.setState({ currentVenue: Number(e.target.value) });
+    await this.props.getRocommendedArtists(this.state.currentVenue);
+    const rec = this.props.artists.filter(
+      artist => artist['recommendations'][0].score <= 9
+    );
+    this.setState({
+      currentBookerRecommandations: rec,
+    });
   };
   render() {
     return (
-      <IonPage>
+      <div className="home">
+        <IonCardHeader className="home" mode="ios">
+          <IonButton
+            mode="ios"
+            href="/artists"
+            className="homeBtn"
+            color="rgb(153, 178, 189);"
+          >
+            Artists
+                      </IonButton>
+          <select onChange={this.handleChange}>
+            {this.props.venues !== undefined &&
+              this.props.venues.map((venue, index) => (
+                <option value={venue.id} key={index}>
+                  {venue.name}
+                </option>
+              ))}
+          </select>
+          <IonCardTitle className="textBox">
+            We got you some artist you might be interested in...
+                      </IonCardTitle>
+        </IonCardHeader>
 
-        <IonHeader >
-          <IonToolbar id="bar" >
-            <IonTitle>Venue location</IonTitle>
-          </IonToolbar>
-        </IonHeader>
+        {this.state.currentBookerRecommandations.map(
+          (artist, index) => {
+            let genres = '';
+            artist['genres'].forEach((el, index) => {
+              genres += el + ' ';
+            });
 
-        <IonContent>
-          <IonCard className="welcome-card">
+            return (
+              <IonCard
+                key={index}
+                href={`/allArtists/${artist['id']}`}
+                className="profile"
+                style={{ width: '250px' }}
+                mode="ios"
+              >
+                <div className="artistBox">
+                  <img src={artist['imageUrl']} alt="img.jpg" />
 
-            <PlacesAutocomplete
-              value={this.state.address}
-              onChange={this.handleChange}
-              onSelect={this.handleSelect}
-            >
-              {({ getInputProps, suggestions, getSuggestionItemProps, loading }) => (
-                <div>
-                  <input
-                    {...getInputProps({
-                      placeholder: 'Search Venues ...',
-                      className: 'searchInput',
-                    })}
-
-                  />
-                  <div className="searchInput">
-                    {loading && <div>Loading...</div>}
-                    {suggestions.map(suggestion => {
-                      const className = suggestion.active
-                        ? 'suggestion-item--active'
-                        : 'suggestion-item';
-
-                      const style = suggestion.active
-                        ? { backgroundColor: '#fafafa', cursor: 'pointer' }
-                        : { backgroundColor: '#ffffff', cursor: 'pointer' };
-                      return (
-                        <div
-                          {...getSuggestionItemProps(suggestion, {
-                            className,
-                            style,
-                          })}
-                        >
-                          <span>{suggestion.description}</span>
-                        </div>
-                      );
-                    })}
-                  </div>
+                  <IonItemGroup style={{ margin: '20px' }}>
+                    <IonCardTitle
+                      style={{ textAlign: 'center' }}
+                      className="artistBoxText"
+                    >
+                      {artist['artistName']}
+                    </IonCardTitle>
+                    <IonCardSubtitle
+                      style={{ textAlign: 'center' }}
+                    >
+                      {genres}
+                    </IonCardSubtitle>
+                  </IonItemGroup>
                 </div>
-              )}
-            </PlacesAutocomplete>
+              </IonCard>
+            );
+          }
+        )}
+      </div>
 
-          </IonCard>
-        </IonContent>
-      </IonPage>
     )
   }
-};
-const mapStateToProps = (state) => ({
-  booker: state.booker
-})
-const mapDispatchToProps = (dispatch) => ({
-  updateVenue: (data) => dispatch(updatedVenue(data))
-})
-export default connect(mapStateToProps, mapDispatchToProps)(Login);
+}
+
+const mapStateToProps = state => ({
+  error: state.user.error,
+  user: state.user,
+  isSearchBarOpen: state.filter.isSearchBarOpen,
+  venues: state.booker.venues,
+  artists: state.artist.allArtists,
+});
+const mapDispatchToProps = dispatch => ({
+  auth: (email, password) => dispatch(auth(email, password)),
+  me: () => dispatch(me()),
+  searchBarValue: value => dispatch(searchBarValue(value)),
+  getRecommendedVenues: id => dispatch(getRecommendedVenues(id)),
+  getRocommendedArtists: id => dispatch(getRocommendedArtists(id)),
+  fetchVenues: id => dispatch(getOneBooker(id)),
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(ArtistRecommendation);
